backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9

sqlalchemy==2.0.32
psycopg[binary]==3.2.1
alembic==1.13.2

passlib[bcrypt]==1.7.4
python-jose==3.3.0

redis==5.0.8
rq==1.16.2

stripe==10.12.0
